Added test cases

+++ contracts/Token.sol
@@ -14,7 +14,7 @@ contract Token {
     function by the compiler.
     */
     /// total amount of tokens
-    uint256 public totalSupply;
+    uint public totalSupply;
 
     /// @param _owner The address from which the balance will be retrieved
     /// @return The balance
